Moved existing .json files from util folder to new i18n/en folder
